Ecommerce-Application-main/backend_python/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
Werkzeug==2.2.2
zipp==3.13.0
